Formatting and install fortran modules
Dummy doxyfile

+++ testing/adios2/bindings/fortran/TestBPWriteTypes_nompi.f90
@@ -3,7 +3,7 @@ program TestBPWriteTypes
     use adios2
     implicit none
 
-    integer, dimension(1) :: shape_dims, start_dims, count_dims
+    integer(kind=8), dimension(1) :: shape_dims, start_dims, count_dims
     integer :: inx, ierr, i
 
     integer(kind=8) :: adios, io, engine
